Split shared and internal tables for behavior plotting.
